api/requirements.txt
fastapi
uvicorn
python-dotenv
supabase
openai
azure-identity
langchain
langchain-openai
